requirements.txt
moderngl>=5.8.2
pygame>=2.1.0
numpy>=1.21.0
PyGLM>=2.5.0
numba>=0.57.0
